extract get_tentative_policy function
