+++ mcts/tree.py
@@ -15,6 +15,7 @@ from board.stone import Stone
 from common.print_console import print_err
 from nn.feature import generate_input_planes
 from nn.network.dual_net import DualNet
+from nn.tentative_policy import get_tentative_policy
 from mcts.batch_data import BatchQueue
 from mcts.constant import NOT_EXPANDED, PLAYOUTS, NN_BATCH_SIZE, \
     MAX_CONSIDERED_NODES, RESIGN_THRESHOLD, MCTS_TREE_SIZE
@@ -504,16 +505,3 @@ class MCTSTree: # pylint: disable=R0902
             "to_move": 'black' if self.to_move == Stone.BLACK else 'white',
         }
         return state
-
-
-def get_tentative_policy(candidates: List[int]) -> Dict[int, float]:
-    """ニューラルネットワークの計算が行われるまでに使用するPolicyを取得する。
-
-    Args:
-        candidates (List[int]): パスを含む候補手のリスト。
-
-    Returns:
-        Dict[int, float]: 候補手の座標とPolicyの値のマップ。
-    """
-    score = np.random.dirichlet(alpha=np.ones(len(candidates)))
-    return dict(zip(candidates, score))
