+++ mcts/tree_async.py
@@ -29,6 +29,7 @@ from mcts.sequential_halving import get_candidates_and_visit_pairs
 from mcts.time_manager import TimeControl, TimeManager
 from nn.feature import generate_input_planes
 from nn.network.dual_net import DualNet
+from nn.tentative_policy import get_tentative_policy
 
 
 class MCTSTreeAsync:  # pylint: disable=R0902
@@ -524,16 +525,3 @@ class MCTSTreeAsync:  # pylint: disable=R0902
             "to_move": "black" if self.to_move == Stone.BLACK else "white",
         }
         return state
-
-
-def get_tentative_policy(candidates: List[int]) -> Dict[int, float]:
-    """ニューラルネットワークの計算が行われるまでに使用するPolicyを取得する。
-
-    Args:
-        candidates (List[int]): パスを含む候補手のリスト。
-
-    Returns:
-        Dict[int, float]: 候補手の座標とPolicyの値のマップ。
-    """
-    score = np.random.dirichlet(alpha=np.ones(len(candidates)))
-    return dict(zip(candidates, score))
